Updated Map class to use proper iteration
There is no reason to use a CopyOnWriteArrayList - it is a lot slower and less efficient. Simply using proper iteration stops CME's from occurring and is much better practice.

In the case where thread safety is also needed, Collections.synchronizedList is a much better alternative for this list as it mutates often.